fix table db fields structure

+++ web/database/migrations/2020_06_29_161822_create_prices_table.php
@@ -17,10 +17,10 @@ class CreatePricesTable extends Migration
             $table->uuid('id')->primary();
 
             $table->integer('stage')->default(0);
-            $table->double('price', 9, 6, true);
+            $table->decimal('price', 10, 6, true);
             $table->unsignedBigInteger('amount');
             $table->unsignedTinyInteger('period_in_days')->default(0);
-            $table->double('percent_profit', 8, 4, true)->default(0);
+            $table->decimal('percent_profit', 8, 3, true)->default(0);
 
             $table->foreignUuid('product_id')->constrained()
                 ->onUpdate('cascade')
